src/utils/data_preprocessor.py
import numpy as np
import torch
from typing import List, Dict, Any
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import DBSCAN

class DataPreprocessor:
    """Industrial Protocol Data Preprocessor"""
    
    def __init__(self):
        self.scaler = StandardScaler()
        self.cluster_model = DBSCAN(eps=0.5, min_samples=5)
        self.protocol_formats = {}
        
    def preprocess_protocol_data(self, raw_packets: List[bytes], 
                               protocol_type: str) -> Dict[str, Any]:
        """Preprocess protocol data"""
        
        # 1. Protocol data clustering and sample expansion
        clustered_data = self._cluster_packets_by_length(raw_packets)
        
        # 2. Data format conversion and encoding standardization
        normalized_data = self._normalize_encoding(clustered_data)
        
        # 3. Data normalization
        preprocessed_data = self._normalize_data(normalized_data)
        
        return {
            'clustered_data': clustered_data,
            'normalized_data': normalized_data,
            'preprocessed_data': preprocessed_data,
            'feature_vectors': self._extract_features(preprocessed_data)
        }
    
    def _cluster_packets_by_length(self, packets: List[bytes]) -> Dict[int, List[bytes]]:
        """Cluster packets by length"""
        length_groups = {}
        
        for packet in packets:
            length = len(packet)
            if length not in length_groups:
                length_groups[length] = []
            length_groups[length].append(packet)
            
        return length_groups
    
    def _normalize_encoding(self, clustered_data: Dict[int, List[bytes]]) -> List[np.ndarray]:
        """Standardize encoding format (convert to 8-bit binary sequences)"""
        normalized_packets = []
        
        for length_group in clustered_data.values():
            for packet in length_group:
                # Convert to binary stream, then to 8-bit binary sequences
                binary_stream = ''.join(format(byte, '08b') for byte in packet)
                
                # Convert to 8-dimensional one-hot vectors (simplified implementation)
                one_hot_vectors = []
                for byte in packet:
                    # Encode each character as 8-dimensional one-hot vector
                    one_hot = np.zeros(8)
                    one_hot[byte % 8] = 1  # Simplified implementation
                    one_hot_vectors.append(one_hot)
                    
                normalized_packets.append(np.array(one_hot_vectors).flatten())
                
        return normalized_packets
    
    def _normalize_data(self, normalized_data: List[np.ndarray]) -> np.ndarray:
        """Data normalization"""
        if not normalized_data:
            return np.array([])
            
        # Find maximum length for padding
        max_length = max(len(vec) for vec in normalized_data)
        padded_data = []
        
        for vec in normalized_data:
            if len(vec) < max_length:
                # Use mixed approach of byte repetition and random padding
                padding_length = max_length - len(vec)
                padding = np.random.choice([0, 1], size=padding_length)
                padded_vec = np.concatenate([vec, padding])
            else:
                padded_vec = vec[:max_length]
                
            padded_data.append(padded_vec)
            
        return np.array(padded_data)
    
    def _extract_features(self, preprocessed_data: np.ndarray) -> np.ndarray:
        """Extract feature vectors"""
        if preprocessed_data.size == 0:
            return np.array([])
            
        # Use standardizer
        if hasattr(self.scaler, 'n_samples_seen_') and self.scaler.n_samples_seen_ > 0:
            features = self.scaler.transform(preprocessed_data)
        else:
            features = self.scaler.fit_transform(preprocessed_data)
            
        return features
    
    def apply_smote_augmentation(self, data: np.ndarray, labels: np.ndarray) -> tuple:
        """Apply SMOTE oversampling technique to enhance minority class data"""
        # Simplified implementation, actual implementation should use imbalanced-learn library
        unique_labels, counts = np.unique(labels, return_counts=True)
        max_count = np.max(counts)
        
        augmented_data = [data]
        augmented_labels = [labels]
        
        for label, count in zip(unique_labels, counts):
            if count < max_count:
                # Duplicate minority class samples to achieve balance
                minority_indices = np.where(labels == label)[0]
                replication_factor = max_count // count
                
                for _ in range(replication_factor - 1):
                    replicated_data = data[minority_indices].copy()
                    # Add small amount of noise
                    noise = np.random.normal(0, 0.01, replicated_data.shape)
                    replicated_data += noise
                    
                    augmented_data.append(replicated_data)
                    augmented_labels.append(np.full(len(minority_indices), label))
        
        return (np.vstack(augmented_data), 
                np.concatenate(augmented_labels))