requirements.txt
# OmniFuzz Dependencies List
# Multi-Agent Reinforcement Learning Based Power IoT Device Protocol-Aware Fuzzing Framework

# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.5.0

# Industrial Protocol Support
pymodbus>=3.0.0
pyshark>=0.5.0
scapy>=2.4.5

# Data Processing and Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and Logging
pyyaml>=6.0
tqdm>=4.62.0

# Network and System Monitoring
psutil>=5.8.0
netaddr>=0.8.0
python-nmap>=0.7.1

# Testing Framework
pytest>=6.2.0
pytest-cov>=2.12.0

# Code Quality
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Documentation Generation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Performance Analysis
tensorboard>=2.8.0
wandb>=0.12.0

# Other Tools
click>=8.0.0
rich>=10.0.0
